+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added tile display.
 - Added ability to add tiles.
 - Added ability to add columns.
+- Added ability to remove columns and tiles in editable-grid.
 - Added redux mock store.
 
 ### Changed